pyproject.toml
[project]
name = "val-quiz"
version = "0.1.0"
description = "a test project for valearnis quiz platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.1.5",
    "django-debug-toolbar>=5.0.1",
    "django-extensions>=3.2.3",
    "django-filter>=24.3",
    "djangorestframework>=3.15.2",
    "markdown>=3.7",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-django>=4.9.0",
]
